+++ src/services/motion.ts
@@ -1,2 +1,3 @@
 export type { Motion } from './motion/Motion';
+export type { MotionConstraints } from './motion/types';
 export { MotionClient } from './motion/Client';
